+++ DataDefinitions/Properties/Ship.Designer.cs
@@ -19,7 +19,7 @@ namespace EddiDataDefinitions.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "18.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     public class Ship {
@@ -147,6 +147,15 @@ namespace EddiDataDefinitions.Properties {
         public static string yourBeluga {
             get {
                 return ResourceManager.GetString("yourBeluga", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to your.
+        /// </summary>
+        public static string yourCaspianExplorer {
+            get {
+                return ResourceManager.GetString("yourCaspianExplorer", resourceCulture);
             }
         }